+++ opensrp-reveal/src/main/java/org/smartregister/reveal/model/FamilyProfileModel.java
@@ -1,5 +1,7 @@
 package org.smartregister.reveal.model;
 
+import org.json.JSONArray;
+import org.json.JSONException;
 import org.json.JSONObject;
 import org.smartregister.commonregistry.CommonPersonObject;
 import org.smartregister.family.domain.FamilyEventClient;
@@ -7,10 +9,17 @@ import org.smartregister.family.model.BaseFamilyProfileModel;
 import org.smartregister.family.util.JsonFormUtils;
 import org.smartregister.family.util.Utils;
 import org.smartregister.reveal.BuildConfig;
+import org.smartregister.reveal.application.RevealApplication;
+import org.smartregister.reveal.util.AppExecutors;
 import org.smartregister.reveal.util.Constants;
+import org.smartregister.reveal.util.Country;
+import org.smartregister.reveal.util.FamilyConstants;
 import org.smartregister.reveal.util.PreferencesUtil;
 import org.smartregister.util.FormUtils;
 
+import timber.log.Timber;
+
+import static org.smartregister.reveal.util.Constants.DatabaseKeys.EVENT_TYPE_FIELD;
 import static org.smartregister.reveal.util.Constants.DatabaseKeys.LAST_NAME;
 import static org.smartregister.reveal.util.FamilyConstants.RELATIONSHIP.RESIDENCE;
 
@@ -36,6 +45,7 @@ public class FamilyProfileModel extends BaseFamilyProfileModel {
     @Override
     public FamilyEventClient processMemberRegistration(String jsonString, String familyBaseEntityId) {
         eventClient = super.processMemberRegistration(jsonString, familyBaseEntityId);
+        setPlanAndTaskIdentifiers(familyBaseEntityId);
         tagEventClientDetails(eventClient);
         return eventClient;
     }
@@ -43,6 +53,9 @@ public class FamilyProfileModel extends BaseFamilyProfileModel {
     @Override
     public FamilyEventClient processFamilyRegistrationForm(String jsonString, String familyBaseEntityId) {
         eventClient = super.processFamilyRegistrationForm(jsonString, familyBaseEntityId);
+        if(org.smartregister.reveal.util.Utils.isCountryBuild(Country.NIGERIA)){
+            setPlanAndTaskIdentifiers(familyBaseEntityId);
+        }
         tagEventClientDetails(eventClient);
         return eventClient;
     }
@@ -50,6 +63,9 @@ public class FamilyProfileModel extends BaseFamilyProfileModel {
     @Override
     public FamilyEventClient processUpdateMemberRegistration(String jsonString, String familyBaseEntityId) {
         eventClient = super.processUpdateMemberRegistration(jsonString, familyBaseEntityId);
+        if(org.smartregister.reveal.util.Utils.isCountryBuild(Country.NIGERIA)){
+            setPlanAndTaskIdentifiers(familyBaseEntityId);
+        }
         tagEventClientDetails(eventClient);
         return eventClient;
     }
@@ -63,11 +79,12 @@ public class FamilyProfileModel extends BaseFamilyProfileModel {
         }
         eventClient.getEvent().addDetails(Constants.Properties.APP_VERSION_NAME, BuildConfig.VERSION_NAME);
         eventClient.getEvent().setLocationId(org.smartregister.reveal.util.Utils.getOperationalAreaLocation(PreferencesUtil.getInstance().getCurrentOperationalArea()).getId());
-        if(planIdentifier != null){
-            eventClient.getEvent().addDetails(Constants.Properties.PLAN_IDENTIFIER,planIdentifier);
+
+        if(getPlanIdentifier() != null){
+            eventClient.getEvent().addDetails(Constants.Properties.PLAN_IDENTIFIER,getPlanIdentifier());
         }
-        if(taskIdentifier !=null){
-            eventClient.getEvent().addDetails(Constants.Properties.TASK_IDENTIFIER,taskIdentifier);
+        if(getTaskIdentifier() !=null){
+            eventClient.getEvent().addDetails(Constants.Properties.TASK_IDENTIFIER,getTaskIdentifier());
         }
     }
 
@@ -113,5 +130,34 @@ public class FamilyProfileModel extends BaseFamilyProfileModel {
 
     public CommonPersonObject getFamilyHeadPersonObject() {
         return familyHeadPersonObject;
+    }
+
+    private void setPlanAndTaskIdentifiers(String familyBaseEntityId) {
+            JSONObject eventsByBaseEntityId = RevealApplication.getInstance().getContext().getEventClientRepository().getEventsByBaseEntityId(familyBaseEntityId);
+            JSONArray events = eventsByBaseEntityId.optJSONArray("events");
+            JSONObject familyRegistrationEvent = null;
+            for (int i = 0; i < events.length(); i++) {
+                try {
+                    JSONObject event = events.getJSONObject(i);
+                    String eventType = event.getString(EVENT_TYPE_FIELD);
+                    if (eventType.equals(FamilyConstants.EventType.FAMILY_REGISTRATION)) {
+                        familyRegistrationEvent = event;
+                        break;
+                    }
+                } catch (JSONException e) {
+                    Timber.e(e);
+                }
+            }
+            if (familyRegistrationEvent != null) {
+                try {
+                    JSONObject details = familyRegistrationEvent.getJSONObject(Constants.DETAILS);
+                    String planIdentifier = details.getString(Constants.Properties.PLAN_IDENTIFIER);
+                    String taskIdentifier = details.getString(Constants.Properties.TASK_IDENTIFIER);
+                    setPlanIdentifier(planIdentifier);
+                    setTaskIdentifier(taskIdentifier);
+                } catch (JSONException e) {
+                    Timber.e(e);
+                }
+            }
     }
 }
